+++ age/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to Rust's notion of
 to 1.0.0 are beta releases.
 
 ## [Unreleased]
+### Added
+- `impl Eq for age::ssh::{ParseRecipientKeyError, UnsupportedKey}`
+
+### Changed
+- MSRV is now 1.65.0.
 
 ## [0.9.2] - 2023-06-12
 ### Added